Documented code 04/12/2021 | 22w15a1 | 0.7.120

+++ TRAP/src/Utils/Hash/ConvertHashToString.h
@@ -15,7 +15,7 @@ namespace TRAP::Utils::Hash
 		for (const uint8_t& i : hash)
 		{
 			ss << std::hex << std::setfill('0') << std::setw(2) << std::nouppercase
-			   << static_cast<int>(static_cast<uint8_t>(i));
+			   << static_cast<int32_t>(i);
 		}
 
 		return ss.str();
